Food4All
Food 4 All is a group project developped at Le Wagon Montréal. 

My team and I have been working on a food sharing app. We want to connect people who experience food insecurity with people in their neighborhood offering free home-cooked meals or grocery store ingredients.

In addition to the human aspect, our application also helps to fight food waste. With the inflation of grocery prices that we are seeing at the moment, this kind of application is really important to have around.